+++ lib/openmls.dart
@@ -43,3 +43,5 @@ export 'src/rust/api/credential.dart';
 export 'src/rust/api/keys.dart';
 export 'src/rust/api/provider.dart';
 export 'src/rust/api/types.dart';
+export 'src/security/secure_bytes.dart';
+export 'src/security/secure_uint8list.dart';